loanfy/requirements.txt
coverage==7.6.1
Django==5.0.8
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-api-response==0.0.30
djangorestframework-simplejwt==5.3.1
drf-simple-apikey==2.1.0
django-environ==0.11.2
drf_spectacular==0.27.2
psycopg2-binary==2.9.9
requests==2.32.3



